+++ Runtime/HapticManager.cs
@@ -3,16 +3,31 @@
 * ​
 */
 
+using System.Collections;
+using UnityEditor;
+using UnityEngine;
+
 namespace Interhaptics
 {
 
     [UnityEngine.AddComponentMenu("Interhaptics/HapticManager")]
     public sealed class HapticManager : Internal.Singleton<HapticManager>
     {
+		[SerializeField] private bool debugSwitch = true;
 
-        public void Init()
+		public bool DebugSwitch
+		{
+			get => debugSwitch;
+			set => debugSwitch = value;
+		}
+        [HideInInspector]
+        public bool monoScriptingBackend = false;
+		private bool isRendering = false;
+
+		public void Init()
         {
-            Core.HAR.Init();
+
+			Core.HAR.Init();
             Internal.HapticDeviceManager.DeviceInitLoop();
 #if UNITY_ANDROID && !ENABLE_METAQUEST
             Platforms.Mobile.GenericAndroidHapticAbstraction.Initialize();
@@ -26,25 +41,44 @@ namespace Interhaptics
 #if !UNITY_EDITOR_OSX
             Init();
 #endif
+
         }
 
-        private void LateUpdate()
+        private void Start()
         {
-            //Compute all haptics event
-            Core.HAR.ComputeAllEvents(UnityEngine.Time.realtimeSinceStartup);
-            //Insert device rendering loop here
-            Internal.HapticDeviceManager.DeviceRenderLoop();
+            isRendering = true;
+            StartCoroutine(ComputeLoop());
+        }
+
+        public IEnumerator ComputeLoop()
+        {
+            while (isRendering)
+            {
+                //Compute all haptics event
+                Core.HAR.ComputeAllEvents(UnityEngine.Time.realtimeSinceStartup);
+                //Insert device rendering loop here
+                Internal.HapticDeviceManager.DeviceRenderLoop();
+                yield return new WaitForSeconds(0.030f);
+            }
         }
 
         private void OnApplicationFocus(bool hasFocus)
         {
-            // TODO pause the haptic playback
-        }
+			// TODO pause the haptic playback
+			if (!hasFocus)
+			{
+				Core.HAR.ClearActiveEvents();
+			}
+		}
 
         private void OnApplicationPause(bool pauseStatus)
         {
-            // TODO pause the haptic playback
-        }
+			// TODO pause the haptic playback
+            if (pauseStatus)
+            {
+				Core.HAR.ClearActiveEvents();
+			}
+		}
 
         protected override void OnOnApplicationQuit()
         {
@@ -55,7 +89,6 @@ namespace Interhaptics
             Core.HAR.Quit();
 #endif
         }
-
-    }
+	}
 
 }