+++ Runtime/Platforms/MetaQuest/MetaQuestProvider.cs
@@ -61,7 +61,10 @@ namespace Interhaptics.Platforms.XR
 
             Core.HAR.AddBodyPart(Perception.Vibration, BodyPartID.Bp_Left_palm, 1, 1, 1, 500, false, false, false, true);
             Core.HAR.AddBodyPart(Perception.Vibration, BodyPartID.Bp_Right_palm, 1, 1, 1, 500, false, false, false, true);
-            UnityEngine.Debug.Log("Meta Quest haptic provider initialised");
+            if (HapticManager.Instance.DebugSwitch)
+            {
+            UnityEngine.Debug.Log("Meta Quest haptic provider started.");
+            }
             return true;
         }