+++ Runtime/Platforms/Mobile/AndroidProvider.cs
@@ -55,9 +55,18 @@ namespace Interhaptics.Platforms.Android
         [UnityEngine.Scripting.Preserve]
         public bool Init()
         {
+        #if !ENABLE_IL2CPP && UNITY_ANDROID && !UNITY_EDITOR
+        UnityEngine.Debug.LogError("Interhaptics requires IL2CPP scripting backend for Android. Please change it in Player Settings. Haptics will not play on the Mono scripting backend on the Android platform." + "Source: Android Provider");
+        HapticManager.Instance.monoScriptingBackend = true;
+        return false;
+        #endif
+
             Interhaptics.Platforms.Mobile.GenericAndroidHapticAbstraction.Initialize();
             Core.HAR.AddBodyPart(Perception.Vibration, Hand, 1, 1, 1, SAMPLERATE, false, false, false, false);
-            UnityEngine.Debug.Log("Android haptic provider initialised");
+            if (HapticManager.Instance.DebugSwitch)
+			{
+            UnityEngine.Debug.Log("Android haptic provider started.");
+            }
             return true;
         }