+++ Runtime/Samples/HapticControlGUI.cs
@@ -1,0 +1,55 @@
+﻿/* ​
+* Copyright (c) 2023 Go Touch VR SAS. All rights reserved. ​
+* ​
+*/
+
+using UnityEngine;
+using Interhaptics.Internal;
+
+namespace Interhaptics.Samples
+{
+	public class HapticControlGUI : MonoBehaviour
+	{
+		[SerializeField]
+		private EventHapticManagerMobile eventHapticManagerMobile;
+		[SerializeField]
+		private EventHapticSource eventHapticSource;
+		[SerializeField]
+		private GameObject buttonPlay;
+		[SerializeField]
+		private GameObject buttonStop;
+
+		void Start()
+		{
+			buttonPlay.SetActive(true);
+			buttonStop.SetActive(false);
+		}
+
+		public void StartHaptics()
+		{
+			eventHapticManagerMobile.StopPlayingEventHapticSources();
+			eventHapticSource.PlayEventVibration();
+		}
+		public void StopHaptics()
+		{
+			eventHapticSource.Stop();
+			Interhaptics.Core.HAR.ClearActiveEvents();
+		}
+
+		void Update()
+		{
+			if (eventHapticSource == null) return;
+			if (eventHapticSource.isPlaying)
+			{
+				buttonPlay.SetActive(false);
+				buttonStop.SetActive(true);
+			}
+			else
+			{
+				buttonPlay.SetActive(true);
+				buttonStop.SetActive(false);
+			}
+		}
+	}
+
+}